Styling, Fonts, etc added
Added styled-components and some basic styling
Added a custom font and also installed react-fonts for fun.

Perhaps should of just used the available library to fetch from pokeapi but it seemed like a good idea to just use fetch

The generationPokemon.js needs to wait for pokemonNames to update after map method is completed, probably using a promise or async/await but to be honest I got a little stuck here.